Updated interface for LNAS and its test
- Moved from functions to class functions
- Changed to inplace functions
