+++ lnas/fmt.py
@@ -42,7 +42,22 @@ class LnasFormat:
                 return False
         return True
 
+    def copy(self) -> LnasFormat:
+        return LnasFormat(
+            version=self.version,
+            geometry=self.geometry.copy(),
+            surfaces={s: arr.copy() for s, arr in self.surfaces.items()},
+        )
+
     def geometry_from_surface(self, surface_name: str) -> LnasGeometry:
+        """Build LNAS geometry from a surface
+
+        Args:
+            surface_name (str): Surface name
+
+        Returns:
+            LnasGeometry: Geometry representing surface
+        """        
         if surface_name not in self.surfaces:
             raise KeyError(
                 f"Unable to find surface named {surface_name}. "
@@ -53,6 +68,31 @@ class LnasFormat:
         triangles = self.geometry.triangles[triangles_idxs].copy()
         vertices = self.geometry.vertices.copy()
         return LnasGeometry(vertices=vertices, triangles=triangles)
+
+    def geometry_from_list_surfaces(self, surfaces_names: list[str]) -> tuple[LnasGeometry, np.ndarray]:
+        """Build geometry from list of surfaces
+
+        Args:
+            surfaces_names (list[str]): List of surfaces names to include
+
+        Returns:
+            tuple[LnasGeometry, np.ndarray]: geometry and the array with the original triangle idxs
+        """
+        triangles_use = np.zeros((len(self.geometry.triangles), ), dtype=bool)
+        for s in surfaces_names:
+            if(s not in self.surfaces):
+                raise KeyError(f"Surface named {s} not in LNAS")
+            s_tri_idxs = self.surfaces[s]
+            triangles_use[s_tri_idxs] = True
+
+        # Index of triangles in original LNAS
+        tri_idxs = np.arange(len(self.geometry.triangles), dtype=np.uint32)
+        tri_idxs = tri_idxs[triangles_use]
+
+        lnas_filtered = self.filter_triangles(triangles_use)
+        lnas_filtered.geometry._full_update()
+
+        return (lnas_filtered.geometry, tri_idxs)
 
     @classmethod
     def from_dct(cls, dct: dict[str, Any]) -> LnasFormat:
@@ -158,3 +198,39 @@ class LnasFormat:
         new_lnas = LnasFormat(version=self.version, geometry=new_geometry, surfaces=new_surfaces)
 
         return new_lnas
+
+    def join(self, lnas_fmts: list[LnasFormat], surfaces_suffixes: list[str] | None):
+        """Join into this LNAS a list of other LNAS
+
+        Args:
+            lnas_fmts (list[LnasFormat]): List of LnasFormat to be combined
+            surfaces_suffixes (list[str] | None, optional): Optional suffix list to add to each lnas. Defaults to None.
+        """
+
+        if len(lnas_fmts) < 1:
+            return
+
+        if surfaces_suffixes is not None:
+            if len(surfaces_suffixes) < len(lnas_fmts):
+                raise ValueError("Less surfaces suffixes than required")
+
+        for i, lnas_fmt in enumerate(lnas_fmts):
+            n_verts, n_tris = len(self.geometry.vertices), len(self.geometry.triangles)
+
+            verts_add = lnas_fmt.geometry.vertices.copy()
+            self.geometry.vertices = np.concatenate((self.geometry.vertices.copy(), verts_add), axis=0)
+
+            tri_add = lnas_fmt.geometry.triangles + n_verts
+            self.geometry.triangles = np.concatenate((self.geometry.triangles.copy(), tri_add), axis=0)
+
+            suffix = surfaces_suffixes[i] if surfaces_suffixes is not None else ""
+            for s, arr in lnas_fmt.surfaces.items():
+                key = s + suffix
+                if key in self.surfaces:
+                    raise KeyError(
+                        f"Surface {s} is already in the list of surfaces, provide a suffix for it"
+                    )
+                self.surfaces[key] = arr + n_tris
+
+        self.geometry._full_update()
+
